pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lambda-test-python"
version = "1.1.0"
description = "Simple Python Lambda function for testing deployment workflows"
authors = [
    {name = "DevOps Team", email = "devops@company.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
keywords = ["lambda", "aws", "serverless", "deployment"]

# Runtime dependencies for realistic Lambda function
dependencies = [
    # Core AWS and data processing dependencies
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    
    # JSON and data manipulation
    "jsonschema>=4.20.0",
    "PyYAML>=6.0.1",
    
    # HTTP and API utilities - compatible with boto3
    "urllib3>=1.26.0,<2.0.0",
    "certifi>=2023.11.17",
    
    # Logging and monitoring
    "structlog>=23.2.0",
    
    # Utility libraries
    "pytz>=2023.3",
    
    # Additional realistic dependencies
    "Pillow>=10.0.0",  # Image processing
    "pandas>=2.0.0",   # Data manipulation (large dependency)
    "numpy>=1.24.0",   # Numerical computing
]

# Development dependencies
[project.optional-dependencies]
dev = [
    # Linting tools
    "ruff>=0.1.0",     # Fast Python linter
    "flake8>=5.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    
    # Testing tools
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    
    # Type checking
    "mypy>=1.0.0",
    
    # Security
    "bandit>=1.7.0",
    
    # Documentation
    "sphinx>=5.0.0",
]

lint = [
    "flake8>=5.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

# Project URLs
[project.urls]
Homepage = "https://github.com/YourOrg/lambda-test-python"
Repository = "https://github.com/YourOrg/lambda-test-python"
Issues = "https://github.com/YourOrg/lambda-test-python/issues"
Documentation = "https://github.com/YourOrg/lambda-test-python/blob/main/README.md"

# Package discovery
[tool.setuptools.packages.find]
where = ["."]
include = ["lambda_function*"]
exclude = ["tests*", "docs*"]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["lambda_function"]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=lambda_function",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Coverage configuration
[tool.coverage.run]
source = ["lambda_function"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
